e_shelf: Improve auto hide behaviour.
Improve auto hide behaviour. Allow the user to re-enter an
auto-hiding shelf as many times as he/she likes. This fixes the
issue where hovering back onto such a configured shelf causes
the shelf to hide and then animate.

This needs to move to edje. As per the TODO task.