systray - make notes of the freedekstop.org version of protocol
there is an FDO version of this. it seems it's not widely supported
but the org.kde is. at some point we probbably have to move over but
for now there isn't a need, but make note of this and have DOMAIN able
to switch in a heartbeat if we want to.