tiling: prevent potential recursive apply calls
additionally, this ensures that clients that cannot be layouted are
definitly outside the tree. Without applying the window tree again.

With all this tiling can be used quite normally. If you want to know
exactly what is going on, set notify level to info, then tiling tells
you what cannot be tiled.