xsettings: keep the logic "simple".
No need for hook. Wrong use of the API. The widget already does
this when checking various objects.

Set DPI to a default if X application settings enabled else use
the custom DPI.

GTK applications only respect the changes
if they are started when the xsettings is brought up. This is
confusing. Sorry for the noise :(