procstats: window resource module.
The module sends messages which the theme displays. Thus, the
module requires a theme with an implemented edj script to be
useful.

Currently EFL 1.26 includes a basic text label for this...
Usage is recursive...everything under the window is included.

Rudimentary, as usual...please do something pretty...

RFC. Hopefully I didn't break the universe...again!