winlist - optimize layout calcs when you have more windows
initial "how do i size and lay this out" is a fuzzy algo that tries to
stuff all the windows into a single screen with several rows of
windows (in large mode). it has to trade off sizing for a squarish
layout with mu;ltiple rows so does some passes and tries and bisecting
etc. - the problem is each stage goes and does a lot of object changes
re-laying them out and querying them. this is expensive. this does a
row length calc on its own without the objects to save a whole lot of
overhead.

in theory i could actually skip almost all the object stuf and make
more assumptions and reduce the object fiddliong to just an initial
"how much fluff around a window item in the list and how much fluff
around the winlist (like padding/title and so on) and then just do
some raw math (and even flatten into arrays for cache friendliness).
but it's fast enough right now without a lot of changes. can always
revisit this in future.