connman agent msg handling -  be more clean to try avoid possible seg
As this was reported with no reproduction method - like a "one off" I
can't be sure this is it and fixes it, but looking for agent->msg and
it becoming a wild crazy pointer. So either it was left there from
long ago and has since been forgotten bhy asan and is thus wild, or ..
memory was scribbled over with junk. in the latter case... who knows.
in the former, this may fix that.

@fix