TODO++ - add new weather gadget/module as forecasts will die next month
yahoo is shutting down its free weather http api in feb 2021. the
existing forecasts needed:

1. move to ecore_con_url to remove al the manual http and http proxy
handling to simplify code and make it more robust
2. needed to fix the location configuration to selecyt from a list of
places or on a map
3. needed fixes to popup handling so it worked likc clock and didnt
pop-down on mouse out of gadget

add to that now a rewriter of the whole http api to be different ...
may as well add a whole new module/gadget as thats actually less work
to start again.