desks: Assign some default desktop names
By default, desktops should (realistically) be assigned some form of
name so that they are not blank. This Fixes the issue of no Desktop
labels in the Virtual Desktop menu, and also fixes the issue mentioned
in the below ticket where "Show on Specific Desktop" dialog would not
show any type of description for the desktops when trying to place a
shelf.

NB: Uses the format of Desktop X,Y (ie: Desktop 0,1)

Fixes T8838

@fix