efm - sync much more often to avoid hangs/stalls with lots of data
do big ops like big copies of 3 or 5gb files over to an nfs mount or a
slow thumbdrive... the os can happily store it all in RAM. but then
any stats on that file still being flushed stall - definitely on nfs.
this leads to much badness. so sync ever 0.2 sec OR every 512k worth
of writes to stay interactive. it's a big worse for perf and maybe a
bit worse on wear but it's far more usable.

@opt