mixer - properly update default sink and set def sink too in mixer popup
mixer popup SAVED the def mixer but never SET it...  and this led to a
bunch of issues when getting gixed. now its saved and set. NOW the
mixer popup has a quick way to select the output sink (device). handy
if uyou use usb audio for heapdhones but on board audio for speakers
for example (or hdmi audio or something).

@fix