mixer - add icon database for sink icons for mixer popup
this will allow us to have specific icon for different sinks - eg a
screeen for hdmi sinks etc. - need to build a "database" in
sink-icons.txt over time. format is:

glob|icons-name
glob2|icon-name2
...

where glob/glob2 are file globs to match the strng of a sink name and
icon-name/icon-name2 are the naime of the icon in the elm icon theme
for xdg/fdo standard icons or extended icon names we added in the elm
theme. add more lines as needed. first match "wins" so put most
specific matches at the front and least specific at the end.