mod bz5 - fix hw that comes with bt rfkill on and bz adapters empty
so - it seems on some hardware and./or os combinations the list of bt
adapters is empty from bluez5 entirely if the adapter is rfkilled off.
this means e doesnt see it at all - has no idea it's there. bz doesnt
expose it. without that we can't turn it on even. so - if our bt
adapter list is empty, then assume this is an error and manually list
rfkillable devines and forcibly unblock them ot make them appear in
bulez5... not so much a bug fix as a system brokenness workaround with
bz5 just ignoring rfkill bnlocked devices and not even telling us
about them.

@fix