mixer - show a mic/recording slider if someothing is and icons
so now display icons for clients/apps that are playing anything back
right now next to the vu meter. also if some app is recording
something, display a slider for mic/input source volume and a list of
icons that are recording too and a vu meter too for this (95% of the
time things wont record so hide it then and only show when something
does).