ibox - track timer and elete it when icon deleted... bad tracking of tmr
if you create a timer for some object and are passing in that ptr to
the timer later.. what happens if that icon (object, struct etc.) is
deleted BEFORE the timer fires? timer fires passing in a junk ptr to
freed memory... and this causes bad things™. don't do this. track your
timers and delete them...