use XSession type - it seems some dm's are enforcing a non-standard
they insist on Type=XSession or you end up with a blank screen. this
was never defined as the standard. wm's have been using
Type=Application for decades! it's still documented in many places as
the way to add a custom session too. this is a workaround these stupid
DM's insisting on this. There is no good reason to make these changes.