bluez fix vanishing obj if we see iface remove double check via get props
so we see an interfaceremoved from dbus for a bz obj - but it actually
isnt removed. why se see this? i don't know... tha'ts odd. but it's
stil there.. so double check and get properties - if we can without
error the obj is not gone - if there is an error from dbus.. it is
gone.

@fix